# backend/backend/openapi.yaml
openapi: 3.0.3
info:
  title: PetClinic API
  version: 1.0.0
  description: |
    REST API for the Spring PetClinic schema (TypeScript / Express / Sequelize port).

    Interactive docs: `/api/docs`. Raw spec: `/api/docs.json`.

    **Authentication.** Every `/api/*` endpoint requires a Bearer access token
    obtained from `POST /auth/login` or `POST /auth/signup`. Unauthenticated
    requests get `401`. The `/auth/*` endpoints, `/health`, and the docs are
    public.
servers:
  - url: /
    description: Same origin as the running server.
security:
  - bearerAuth: []
...
paths:
...
  /api/owners/{ownerId}/pets/{petId}/visits:
    parameters:
      - $ref: '#/components/parameters/OwnerId'
      - $ref: '#/components/parameters/PetId'
    get:
      tags: [Owners, Visits]
      summary: List visits for a pet belonging to this owner
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    post:
      tags: [Owners, Visits]
      summary: Add a visit to a pet belonging to this owner
      description: '`pet_id` is taken from the URL; do not send it in the body.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VisitInputForPet' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
  /api/vets:
    get:
      tags: [Vets]
      summary: List vets (with specialties included)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Vet' }
    post:
      tags: [Vets]
      summary: Create vet
      description: |
        `specialties`, if provided, replaces the vet's specialty membership.
        Accepts either an array of specialty ids or an array of `{ id }` objects.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VetInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/vets/{vetId}:
    parameters:
      - $ref: '#/components/parameters/VetId'
    get:
      tags: [Vets]
      summary: Get vet by id (with specialties included)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Vets]
      summary: Update vet
      description: |
        If `specialties` is present in the body, the vet's specialty membership is
        replaced wholesale (delete + bulk insert). Omit the field to leave it untouched.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VetInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Vet' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Vets]
      summary: Delete vet (also clears its specialty join rows)
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /api/visits:
    get:
      tags: [Visits]
      summary: List visits (newest first, with pet included)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Visit' }
    post:
      tags: [Visits]
      summary: Create visit
      description: 'Use `pet_id` in the body. For visits scoped to an owner+pet, prefer `POST /api/owners/{ownerId}/pets/{petId}/visits`.'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VisitInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
  /api/visits/{visitId}:
    parameters:
      - $ref: '#/components/parameters/VisitId'
    get:
      tags: [Visits]
      summary: Get visit by id (with pet included)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
    put:
      tags: [Visits]
      summary: Update visit
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VisitInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Visit' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
    delete:
      tags: [Visits]
      summary: Delete visit
      responses:
        '204': { description: No Content }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token issued by `/auth/login`, `/auth/signup`, or `/auth/refresh`.
  parameters:
    OwnerId:
      in: path
      name: ownerId
      required: true
      schema: { type: integer, minimum: 0 }
    PetId:
      in: path
      name: petId
      required: true
      schema: { type: integer, minimum: 0 }
    VetId:
      in: path
      name: vetId
      required: true
      schema: { type: integer, minimum: 0 }
    VisitId:
      in: path
      name: visitId
      required: true
      schema: { type: integer, minimum: 0 }
...
  responses:
    BadRequest:
      description: Malformed request (bad path param, missing required field, validation failure).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: Unique-constraint or foreign-key violation.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Owner with id 99 not found
        details:
          type: array
          items: { type: string }
          description: Per-field validation messages (Sequelize validation / unique errors).
        path:
          type: string
          description: Returned for the global "Route not found" 404.
    Owner:
      type: object
      required: [id]
      properties:
        id: { type: integer, readOnly: true, example: 1 }
        first_name: { type: string, nullable: true, example: George }
        last_name:  { type: string, nullable: true, example: Franklin }
        address:    { type: string, nullable: true, example: 110 W. Liberty St. }
        city:       { type: string, nullable: true, example: Madison }
        telephone:  { type: string, nullable: true, example: '6085551023' }
        pets:
          type: array
          items: { $ref: '#/components/schemas/Pet' }
          description: Present when fetched via the owners endpoints (which include pets, pet types, and visits).
...
    Pet:
      type: object
      required: [id, type_id, owner_id]
      properties:
        id:         { type: integer, readOnly: true, example: 1 }
        name:       { type: string, nullable: true, example: Leo }
        birth_date: { type: string, format: date, nullable: true, example: '2010-09-07' }
        type_id:    { type: integer, example: 1 }
        owner_id:   { type: integer, example: 1 }
        type:
          allOf:
            - $ref: '#/components/schemas/PetType'
          description: Included via Sequelize association.
        owner:
          allOf:
            - $ref: '#/components/schemas/Owner'
          description: Included via Sequelize association.
        visits:
          type: array
          items: { $ref: '#/components/schemas/Visit' }
          description: Included via Sequelize association.
...
    Vet:
      type: object
      required: [id]
      properties:
        id:         { type: integer, readOnly: true }
        first_name: { type: string, nullable: true }
        last_name:  { type: string, nullable: true }
        specialties:
          type: array
          items: { $ref: '#/components/schemas/Specialty' }
    VetInput:
      type: object
      properties:
        first_name: { type: string, nullable: true }
        last_name:  { type: string, nullable: true }
        specialties:
          description: Replaces the vet's specialty membership when present. Accepts plain ids or `{ id }` objects.
          oneOf:
            - type: array
              items: { type: integer }
              example: [1, 2]
            - type: array
              items:
                type: object
                required: [id]
                properties:
                  id: { type: integer }
              example:
                - { id: 1 }
                - { id: 2 }
    Specialty:
      type: object
      required: [id]
      properties:
        id:   { type: integer, readOnly: true }
        name: { type: string, nullable: true }
...
    PetType:
      type: object
      required: [id]
      properties:
        id:   { type: integer, readOnly: true }
        name: { type: string, nullable: true }
...
    Visit:
      type: object
      required: [id, pet_id]
      properties:
        id:          { type: integer, readOnly: true }
        pet_id:      { type: integer }
        visit_date:  { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
        pet:
          allOf:
            - $ref: '#/components/schemas/Pet'
          description: Included via Sequelize association.
    VisitInput:
      type: object
      description: Body for `POST /api/visits` and `PUT /api/visits/{visitId}`.
      required: [pet_id]
      properties:
        pet_id:      { type: integer }
        visit_date:  { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
    VisitInputForPet:
      type: object
      description: Body for visit endpoints nested under a pet. `pet_id` is taken from the URL.
      properties:
        visit_date:  { type: string, format: date, nullable: true }
        description: { type: string, nullable: true }
...